split RunningMean to separate file
